+++ src/oauth.tsx
@@ -1,7 +1,7 @@
 import { base64 } from "@hexagon/base64";
 import { zValidator } from "@hono/zod-validator";
 import { eq } from "drizzle-orm";
-import { Hono } from "hono";
+import { type Context, Hono } from "hono";
 import { cors } from "hono/cors";
 import { createMiddleware } from "hono/factory";
 import { z } from "zod";
@@ -322,7 +322,7 @@ const tokenRequestSchema = z.object({
   code: z.string().optional(),
   client_id: z.string(),
   client_secret: z.string(),
-  redirect_uri: z.string().url(),
+  redirect_uri: z.string().url().optional(),
   scope: scopesSchema.optional(),
 });
 
@@ -383,6 +383,18 @@ app.post("/token", cors(), async (c) => {
         400,
       );
     }
+
+    if (!form.redirect_uri) {
+      return c.json(
+        {
+          error: "invalid_request",
+          error_description:
+            "The authorization code grant flow requires a redirect URI.",
+        },
+        400,
+      );
+    }
+
     const token = await db.query.accessTokens.findFirst({
       where: eq(accessTokens.code, form.code),
       with: { application: true },
@@ -399,6 +411,20 @@ app.post("/token", cors(), async (c) => {
         400,
       );
     }
+
+    // Validate that the redirect URI given is registered with the Application
+    // (since we"re not tracking Access Grants which would bind the redirect URI
+    // to the code)
+    if (!token.application.redirectUris.includes(form.redirect_uri)) {
+      return c.json(
+        {
+          error: "invalid_request",
+          error_description: "Invalid redirect URI.",
+        },
+        400,
+      );
+    }
+
     const now = (Date.now() / 1000) | 0;
     const message = `${now}^${token.code}`;
     const textEncoder = new TextEncoder();
@@ -443,5 +469,27 @@ app.post("/token", cors(), async (c) => {
     created_at: (+tokens[0].created / 1000) | 0,
   });
 });
+
+export async function oauthAuthorizationServer(c: Context) {
+  const url = new URL(c.req.url);
+
+  return c.json({
+    issuer: new URL("/", url).href,
+    authorization_endpoint: new URL("/oauth/authorize", url).href,
+    token_endpoint: new URL("/oauth/token", url).href,
+    // Not yet supported by Hollo:
+    // "revocation_endpoint": "",
+    scopes_supported: scopeEnum.enumValues,
+    response_types_supported: ["code"],
+    response_modes_supported: ["query"],
+    grant_types_supported: ["authorization_code", "client_credentials"],
+    token_endpoint_auth_methods_supported: [
+      "client_secret_post",
+      // Not supported by Hollo:
+      // "client_secret_basic",
+    ],
+    app_registration_endpoint: new URL("/api/v1/apps", url).href,
+  });
+}
 
 export default app;